css/css-animations/parsing/animation-name-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-name with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-name">
<meta name="assert" content="animation-name supports only the grammar '[ none | <keyframes-name> ]#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("animation-name", '12');
test_invalid_value("animation-name", 'one two');

test_invalid_value("animation-name", 'one, initial');
test_invalid_value("animation-name", 'one, inherit');
test_invalid_value("animation-name", 'one, unset');
test_invalid_value("animation-name", 'default, two');
test_invalid_value("animation-name", 'revert, three');
test_invalid_value("animation-name", 'revert-layer, four');

test_invalid_value("animation-name", '""');

// Restricted keywords as strings
test_invalid_value("animation-name", '"none"');
test_invalid_value("animation-name", '"NoNe"');
test_invalid_value("animation-name", '"initial"');
test_invalid_value("animation-name", '"INITIAL"');
test_invalid_value("animation-name", '"inherit"');
test_invalid_value("animation-name", '"INHERIT"');
test_invalid_value("animation-name", '"revert"');
test_invalid_value("animation-name", '"REVERT"');
test_invalid_value("animation-name", '"revert-layer"');
test_invalid_value("animation-name", '"REVERT-LAYER"');
test_invalid_value("animation-name", '"unset"');
test_invalid_value("animation-name", '"UNSET"');
test_invalid_value("animation-name", '"default"');
test_invalid_value("animation-name", '"DEFAULT"');

</script>
</body>
</html>
